#150 [refactor] 소셜로그인 액티비티에서 프래그먼트로 전환
